Move wallentry to object file
