+++ src/openrct2/world/Scenery.h
@@ -96,27 +96,6 @@ enum LARGE_SCENERY_FLAGS
     LARGE_SCENERY_FLAG_HAS_TERTIARY_COLOUR = (1 << 6),  // 0x40
 };
 
-enum WALL_SCENERY_FLAGS
-{
-    WALL_SCENERY_HAS_PRIMARY_COLOUR = (1 << 0),   // 0x1
-    WALL_SCENERY_HAS_GLASS = (1 << 1),            // 0x2
-    WALL_SCENERY_CANT_BUILD_ON_SLOPE = (1 << 2),  // 0x4
-    WALL_SCENERY_IS_DOUBLE_SIDED = (1 << 3),      // 0x8
-    WALL_SCENERY_IS_DOOR = (1 << 4),              // 0x10
-    WALL_SCENERY_LONG_DOOR_ANIMATION = (1 << 5),  // 0x20
-    WALL_SCENERY_HAS_SECONDARY_COLOUR = (1 << 6), // 0x40
-    WALL_SCENERY_HAS_TERTIARY_COLOUR = (1 << 7),  // 0x80
-};
-
-enum WALL_SCENERY_2_FLAGS
-{
-    WALL_SCENERY_2_NO_SELECT_PRIMARY_COLOUR = (1 << 0), // 0x1
-    WALL_SCENERY_2_DOOR_SOUND_MASK = 0x6,
-    WALL_SCENERY_2_DOOR_SOUND_SHIFT = 1,
-    WALL_SCENERY_2_IS_OPAQUE = (1 << 3), // 0x8
-    WALL_SCENERY_2_ANIMATED = (1 << 4),  // 0x10
-};
-
 struct SmallSceneryEntry
 {
     StringId name;
@@ -138,18 +117,6 @@ struct SmallSceneryEntry
     }
 };
 
-struct WallSceneryEntry
-{
-    StringId name;
-    uint32_t image;
-    CursorID tool_id;
-    uint8_t flags;
-    uint8_t height;
-    uint8_t flags2;
-    money16 price;
-    ObjectEntryIndex scenery_tab_id;
-    uint8_t scrolling_mode;
-};
 enum class PathBitDrawType : uint8_t;
 
 struct PathBitEntry
@@ -281,6 +248,8 @@ void SceneryInit();
 void SceneryUpdateTile(const CoordsXY& sceneryPos);
 void ScenerySetDefaultPlacementConfiguration();
 void SceneryRemoveGhostToolPlacement();
+
+struct WallSceneryEntry;
 
 WallSceneryEntry* GetWallEntry(ObjectEntryIndex entryIndex);
 BannerSceneryEntry* GetBannerEntry(ObjectEntryIndex entryIndex);
